refactor: change namespace from many files in repositories folder
