+++ LeaveManagementWebAPI/Repositories/Datas/DepartementTypeRepository.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using System.Threading.Tasks;
 using LeaveManagementWebAPI.Contexts;
 using LeaveManagementWebAPI.Models.ViewModels;
-using LeaveManagementWebAPI.Repositories.Interface;
+using LeaveManagementWebAPI.Repositories.Interfaces;
 using LeaveManagementWebAPI.Models;
 
-namespace LeaveManagementWebAPI.Repositories.Data
+namespace LeaveManagementWebAPI.Repositories.Datas
 {
     public class DepartementTypeRepository : IDepartementTypeRepository
     {
