+++ LeaveManagementWebAPI/Repositories/Interfaces/IDepartementTypeRepository.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using LeaveManagementWebAPI.Models.ViewModels;
 
-namespace LeaveManagementWebAPI.Repositories.Interface
+namespace LeaveManagementWebAPI.Repositories.Interfaces
 {
     interface IDepartementTypeRepository
     {
